oq-ui-client2/src/exposure/exposure_export.js:
Changed the exposure export form load request from a POST to a GET.

We were using jQuery.load() before, but if you pass parameters to it,
it it automatically treated as a POST (lame!). Unfortunately,
jQuery.get() doesn't support partial document loads the way load()
does, so I had to basically add the equivalent code.