static/css:
Massive css cleanup, mostly to the header and footing margins/padding.

Also added a css file just for the index page to let the footer
sink to the bottom of the page, whereas with the map-based apps/pages,
we want to pin the footer to the bottom and allow no scrolling.