icebox/urls:
Added url mappings for `list_artifact_groups` and `get_artifact_group`.

Also adding trailing slashes to view urls, for consistency and to avoid
problems with clients (if you don't have the trailing slash, this can be
really frustrating-- /icebox/foo and /icebox/foo/ would not be the same,
and a client trying to access the first one would not work).

See also the django setting APPEND_SLASH.