static/js/calculate:
Fixed a bug where, if you load a calculation, the dialog gets set such
that the close button (the [x] in the upper right) is hidden. This is
fine for the "Loading..." dialog, but it's a problem because after that
the style gets applied to the "Run calculation" dialogs.

This patch ensures that the styling is reset.