remove the old spectrum curve, and dedicate a custom function for new spectrum curve generation
Previously, the spectrum curves rendering was based closely on the UHS
curve code. However it has been determined that the spectrum curves
must take a different approach because they do not have uniform x-axis
values. Because the UHS curves store the x-axis values in a UTF-grid
metadata field, all curve can share a common x-axis array.
To render spectrum curves, we could pass the x and y axis values for
each location into the UTfGrid. Another approach is to pass into the
Utfgrid a set of 5 parameters that can be used to create the curve for
each location. This approach requires far less data, but instead
requires computation on the client.