load the project list each time the load project button is activated
Rather then on page load loading and saving in memory the project list,
I am changing the logic so that each time the ‘load project’ button is
clicked, the api is called to generate a new project list. In this way
if a new project is uploaded in mid session, the user can access it
without reloading the page. This api call is very fast, so there is no
degradation in performance.