Investigate primary indicator incomplete data issue
It seems that there is a bit more to this bug. Another issue with the
sample data that is causing the primary indicator chart to break is
that in the data set, the primary indicators are not scaled from 0 to 1
as per a requirement dating back some time. I am now asking the project
owner about this constraints and some possible solutions.