Mapbox-GL / Leaflet integration proof of concept
LeafletJs is not able to scale up when a GeoJson containing a large
number of vertices is loaded into a map layer.

As a possible solution, I have incorporated Mapbox-GL into the IRV
application. Mapbox-GL renders the GeoJson as vector tiles.

This implementation requires no server side changes to the OQP, the
data is retrieved from Geoserver just as it was before.

TODO:
* Implement Web-GL browser support checker
* Implement layer filters
* Dynamically update the layer data/style according to SVIR weight
changes