Deep copy the layer attributes obj
This application modifies the layer attributes based on changes to the
indicator weights.
The app also need to keep track of the original layer attributes values.
Thus, the layerAttributes obj contains "properties" and "newProperties".
This presents an issue when using the object as a data source in
Mapbox-GL, because
it is hard coded to use properties. In order to get around this issue,
we need to
deep copy the layerAttributes, and replace "properties" with
newProperties