Refactor the vulnerability function form - part 1
Per todays conversation with the project owners, it has been decided to
change the vulnerability function  workflow. Now we are providing the
user with a single button to create a new function form/table. From
this form, if the user selects a distribution type of LN or BT then the
rendered table with be of type discrete, while PM will render a
probability mass function table.