Build the vulnerability prob mass function PM distribution type XML tags
This is a tricky bit block of code, because the table row and columns
are inverse from previous versions. Because of this, I needed to
‘pre-build’ a custom object with array of each column of data values,
then build the xml probability tags. There is also addition complexity
because the first column represents the Loss ratio, and the first row
represents the IML.